requirements.txt
fastai==2.7.12
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
python-jose==3.3.0
pydantic==2.5.0
numpy==1.26.4
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.18.1
transformers==4.41.0